[C10D] Add record_id to flight recorder
In cases where sequence number is shared between events (e.g. coalesced
collectives) we want to ensure a unique (and ordered) ID per record.

Note: the records are already in a list, so their ID could be implicitly
observed.  But (1) it's a ring buffer, so absolute ID is lost once the
buffer rolls over once, (2) users may sort or process or filter their
flight records, so having the ID be an explicit member of an entry is
still useful

ghstack-source-id: 1cb8cf5d24f6c11b2613c786978ea1499d8c7107
Pull Request resolved: https://github.com/pytorch/pytorch/pull/120724